requirements.txt
Flask~=2.1.2
Pillow~=9.1.1
SQLAlchemy~=1.4.36
opencv-python~=4.5.5.64
numpy~=1.22.4
